Eliminate high CPU usage / battery drain on mobile
Modified to only use the pbLoader class when the user views slideshow. The pbLoader class uses css animation. CPU usage is high when animation is enabled - which was always without this change.
CPU usage and battery consumption (mobile) drops to next to nothing when the user is not viewing the slideshow because constant DOM rendering is eliminated.